moved ReaderSearchService's logic out to ReaderSearchLogic in preparation for code sharing
